+++ ee/sites/src/components/SubmissionListItem.tsx
@@ -1,4 +1,4 @@
-import type { SubmissionListingSiteContext } from '../routes/$siteName.submissions._index/site-context.format.server.js';
+import type { SubmissionListingSiteContext } from '../routes/$siteName.submissions-classic/site-context.format.server.js';
 import { Link } from 'react-router';
 import { clientCheckSiteScopes, formatDate, formatToNow, primitives } from '@curvenote/scms-core';
 import {
@@ -9,7 +9,7 @@ import {
   SubmissionKind,
   Slug,
 } from './Chips.js';
-import type { AugmentedSubmissionListingItem } from '../routes/$siteName.submissions._index/types.js';
+import type { AugmentedSubmissionListingItem } from '../routes/$siteName.submissions-classic/types.js';
 import { SubmissionActionsArea } from './SubmissionActionsArea.js';
 
 export function SubmissionListItem({
