+++ ee/sites/src/routes/$siteName.submissions-classic/CollectionSelect.tsx
